Relocate tests to proper file
